requirements.txt
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.5
cycler==0.11.0
Cython==0.29.28
filelock==3.8.2
fonttools==4.33.3
ftfy==6.1.1
gensim==4.2.0
google-auth==2.8.0
h5py==3.7.0
huggingface-hub==0.11.1
idna==3.3
importlib-metadata==4.11.4
kiwisolver==1.4.3
matplotlib==3.5.3
numpy==1.21.6
oauthlib==3.2.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
PyMySQL==1.0.2
pyparsing==3.0.9
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyqt5-tools==5.15.4.3.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.1
PyYAML==6.0
QDarkStyle==3.1
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
QtPy==2.3.0
regex==2022.10.31
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.7.3
sentencepiece==0.1.97
six==1.16.0
smart-open==6.2.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tokenizers==0.13.2
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.64.1
traitlets==5.3.0
transformers==4.25.1
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
wincertstore==0.2
zipp==3.8.0
